Move input formats to new packages
