+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/ActionProcessor.java
@@ -27,6 +27,8 @@ import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import ca.on.oicr.gsi.shesmu.input.shesmu.ShesmuIntrospectionProcessorRepository;
+import ca.on.oicr.gsi.shesmu.input.shesmu.ShesmuIntrospectionValue;
 import io.prometheus.client.Gauge;
 
 /**
