--- shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/GsiStdRepository.java
@@ -1,5 +1,0 @@
-package ca.on.oicr.gsi.shesmu;
-
-public interface GsiStdRepository extends InputRepository<GsiStdValue> {
-
-}
