+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/input/gsistd/GsiStdRepository.java
@@ -1,0 +1,7 @@
+package ca.on.oicr.gsi.shesmu.input.gsistd;
+
+import ca.on.oicr.gsi.shesmu.InputRepository;
+
+public interface GsiStdRepository extends InputRepository<GsiStdValue> {
+
+}
