+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/input/shesmu/ShesmuIntrospectionRepository.java
@@ -1,4 +1,6 @@
-package ca.on.oicr.gsi.shesmu;
+package ca.on.oicr.gsi.shesmu.input.shesmu;
+
+import ca.on.oicr.gsi.shesmu.InputRepository;
 
 public interface ShesmuIntrospectionRepository extends InputRepository<ShesmuIntrospectionValue> {
